src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
apturl==0.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.8.0
backcall==0.2.0
backoff==2.1.2
bbndb==2020.1
bcrypt==3.2.0
blinker==1.4
bqplot==0.12.36
Brlapi==0.8.3
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
devscripts===2.22.1ubuntu1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
executing==1.0.0
fasteners==0.14.1
filelock==3.6.0
Flask==2.2.2
frozenlist==1.3.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
gpg===1.16.0-unknown
gql==3.4.0
graphql-core==3.2.1
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.15.3
ipython==8.5.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
JPype1==1.4.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
natsort==8.0.2
nest-asyncio==1.5.5
netifaces==0.11.0
networkx==2.8.6
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.5.0
paramiko==2.9.3
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.13.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.0
QGL==2020.1
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
rsa==4.8
scipy==1.9.1
scour==0.38.2
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.5.0
systemd-python==234
tornado==6.2
traitlets==5.4.0
traittypes==0.2.1
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unidiff==0.5.5
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
